[MOV] discuss, rtc: move some RTC related code to a new module
todo:

* migrate `mail.sfu_server_key`, `mail.sfu_server_url` to
`rtc.sfu_server_key`, rtc.sfu_server_key.

* notify `nim` about changes on sfu key file (`mail.tools.discuss` to
`rtc.tools.sfu`) for saas override change.
