+++ addons/mail/__manifest__.py
@@ -59,7 +59,7 @@ For more specific needs, you may also assign custom-defined actions
 (technically: Server Actions) to be triggered for each incoming mail.
     """,
     'website': 'https://www.odoo.com/app/discuss',
-    'depends': ['base', 'base_setup', 'bus', 'web_tour', 'html_editor'],
+    'depends': ['base', 'base_setup', 'bus', 'rtc', 'web_tour', 'html_editor'],
     'data': [
         'data/mail_groups.xml',
         'wizard/mail_activity_schedule_views.xml',
@@ -194,9 +194,6 @@ For more specific needs, you may also assign custom-defined actions
         'web.qunit_suite_tests': [
             'mail/static/tests/legacy/**/*',
             ('remove', 'mail/static/tests/legacy/helpers/**/*'),
-        ],
-        'mail.assets_odoo_sfu': [
-            'mail/static/lib/odoo_sfu/odoo_sfu.js',
         ],
         'mail.assets_lamejs': [
             'mail/static/lib/lame/lame.js',
