+++ addons/mail/models/mail_push_device.py
@@ -5,7 +5,7 @@ import json
 import logging as logger
 
 from odoo import api, fields, models
-from ..tools.jwt import generate_vapid_keys, InvalidVapidError
+from odoo.addons.web.tools.jwt import generate_vapid_keys, InvalidVapidError
 
 _logger = logger.getLogger(__name__)
