+++ addons/mail/views/res_config_settings_views.xml
@@ -80,18 +80,6 @@
                                 </div>
                             </div>
                         </setting>
-                        <setting string="SFU server">
-                            <div class="content-group">
-                                <div class="row mt16">
-                                    <label for="sfu_server_url" string="URL" class="col-lg-3"/>
-                                    <field name="sfu_server_url"/>
-                                </div>
-                                <div class="row mt16">
-                                    <label for="sfu_server_key" string="Key" class="col-lg-3"/>
-                                    <field name="sfu_server_key"/>
-                                </div>
-                            </div>
-                        </setting>
                         <setting string="Custom ICE server list" help="Configure your ICE server list for webRTC">
                             <div class="content-group">
                                 <div class="row col-lg-4">
