+++ addons/rtc/data/neutralize.sql
@@ -1,0 +1,3 @@
+-- * delete SFU keys
+DELETE FROM ir_config_parameter
+    WHERE key IN ('rtc.sfu_server_key');
