+++ addons/rtc/static/lib/odoo_sfu/odoo_sfu.js
