+++ addons/rtc/tools/sfu.py
@@ -1,0 +1,15 @@
+import os
+
+def get_sfu_url(env) -> str | None:
+    sfu_url = env['ir.config_parameter'].sudo().get_param("rtc.sfu_server_url")
+    if not sfu_url:
+        sfu_url = os.getenv("ODOO_SFU_URL")
+    if sfu_url:
+        return sfu_url.rstrip("/")
+
+
+def get_sfu_key(env) -> str | None:
+    sfu_key = env['ir.config_parameter'].sudo().get_param('rtc.sfu_server_key')
+    if not sfu_key:
+        return os.getenv("ODOO_SFU_KEY")
+    return sfu_key
