+++ addons/test_mail/tests/test_mail_push.py
@@ -10,7 +10,7 @@ import odoo
 
 from odoo.tools.misc import mute_logger
 from odoo.addons.mail.tests.common import mail_new_test_user
-from odoo.addons.mail.tools.jwt import InvalidVapidError
+from odoo.addons.web.tools.jwt import InvalidVapidError
 from odoo.addons.sms.tests.common import SMSCommon
 from odoo.addons.test_mail.data.test_mail_data import MAIL_TEMPLATE
 from odoo.tests import tagged
